Theme: Consolidate color warning diagnostics
